Object-oriented rewrite and enable AirBnB linting
My dad enjoyed this animated card but wondered whether it passed lint.
It didn't.

So what began as a 'simple' task of fulfilling the linter's demands led
to the realisation that the code should be substantially restructured,
initially in order to remove the questionable circular dependencies of
branches and leaves that relied on JavaScript's 'hoisting' of
declarations to work.

The tree is now handled through a Tree object. Branches and leaves more
logically inherit from a TreeNode class, which contains common methods.
Additionally, each generation of branches and leaves is its own
Generation object, making the code for animating the branches and leaves
much clearer.

To create an animated tree, it is now only necessary to instantiate the
Tree class, and drawtree.js is now just a wrapper that connects the
classes up to elements in the page and sets parameters.

The parameters object has also been modified, containing sub-objects for
parameters of branches and leaves. This has required some rather clunky
functions for getting and setting parameters which are currently
global, but a parameters class would be a better implementation.