Fix timing bugs
A mistake meant that two instances of the root node would be created,
one on top of the other, resulting in the appearance of a pause in the
animation. This is fixed.

Additionally, nonlinear easing is applied to the growth of the
root node, and timings are adjusted accordingly. Responsibility for
timing is moved to the Generation object, not individual nodes.